proxy/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.9.0
openai==1.50.0
google-generativeai==0.8.0
anthropic==0.34.0
